workaround for quic packet number parsing issue
There is some egregious packet reordering happening at the lokinet level which causes
the QUIC packet number encoding to be interpreted incorrectly, i.e. it thinks it has received
a packet from the future, sends and ACK for that, and the other end (rightly) complains.

The packet ordering issue needs to be resolved, but this 'fixes' the symptom in the meantime.